src/main/java/guru/springframework/recipe/app/converters/fromcommand/UnitOfMeasureCommandToUnitOfMeasure.java
package guru.springframework.recipe.app.converters.fromcommand;

import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

import guru.springframework.recipe.app.commands.UnitOfMeasureCommand;
import guru.springframework.recipe.app.domain.UnitOfMeasure;
import lombok.Synchronized;

@Component
public class UnitOfMeasureCommandToUnitOfMeasure implements Converter<UnitOfMeasureCommand, UnitOfMeasure> {

	@Synchronized
	@Nullable
	@Override
	public UnitOfMeasure convert(UnitOfMeasureCommand source) {
		if (source == null) {
			return null;
		}
		
		final UnitOfMeasure destination = new UnitOfMeasure();
		destination.setId(source.getId());
		destination.setDescription(source.getDescription());
		
		return destination;
	}

}
